Add pkg/utils for low-level utilities
pkg/utils is for low-level utility functions used by both
controllers and CloudClient interface implementations.
