+++ pkg/cloudclient/aws/aws.go
@@ -22,6 +22,7 @@ import (
 	configv1 "github.com/openshift/api/config/v1"
 	cloudingressv1alpha1 "github.com/openshift/cloud-ingress-operator/pkg/apis/cloudingress/v1alpha1"
 	"github.com/openshift/cloud-ingress-operator/pkg/config"
+	baseutils "github.com/openshift/cloud-ingress-operator/pkg/utils"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -138,7 +139,7 @@ func readClusterRegionFromConfigMap(kclient client.Client) (string, error) {
 
 //GetClusterRegion returns the installed cluster's AWS region
 func GetClusterRegion(kclient client.Client) (string, error) {
-	infra, err := getInfrastructureObject(kclient)
+	infra, err := baseutils.GetInfrastructureObject(kclient)
 	if err != nil {
 		return "", err
 	} else if infra.Status.PlatformStatus == nil {
