+++ pkg/controller/publishingstrategy/publishingstrategy_controller.go
@@ -12,7 +12,7 @@ import (
 	operatorv1 "github.com/openshift/api/operator/v1"
 	cloudingressv1alpha1 "github.com/openshift/cloud-ingress-operator/pkg/apis/cloudingress/v1alpha1"
 	"github.com/openshift/cloud-ingress-operator/pkg/cloudclient"
-	utils "github.com/openshift/cloud-ingress-operator/pkg/controller/utils"
+	baseutils "github.com/openshift/cloud-ingress-operator/pkg/utils"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
@@ -179,7 +179,7 @@ func (r *ReconcilePublishingStrategy) Reconcile(request reconcile.Request) (reco
 		return reconcile.Result{Requeue: true, RequeueAfter: 10 * time.Second}, nil
 	}
 
-	cloudPlatform, err := utils.GetPlatformType(r.client)
+	cloudPlatform, err := baseutils.GetPlatformType(r.client)
 	if err != nil {
 		log.Error(err, "Failed to create a Cloud Client")
 		return reconcile.Result{}, err
